@@ -28,13 +28,19 @@
     "prepare:help": "echo 'Install (or reinstall) git hooks'",
     "prepare": "if test -d \"../.git\"; then cd .. && husky install web/.husky && rm -rf .git/hooks && ln -s ../web/.husky ./.git/hooks; fi"
   },
+  "overrides": {
+    "@types/react": "18.0.14",
+    "@types/react-dom": "18.0.5",
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
+  },
   "dependencies": {
     "@types/compression": "^1.7.2",
     "@types/dedent": "^0.7.0",
     "@types/morgan": "^1.9.3",
     "@types/node": "^16.11.0",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
+    "@types/react": "18.0.14",
+    "@types/react-dom": "18.0.5",
     "@types/styletron-engine-atomic": "^1.1.1",
     "@types/styletron-engine-monolithic": "^0.0.0",
     "@types/styletron-react": "^5.0.3",
@@ -56,14 +62,15 @@
     "express": "^4.18.0",
     "helmet": "^5.0.2",
     "husky": "^7.0.2",
+    "ky": "^0.30.0",
     "lint-staged": "^11.2.3",
     "lucide-react": "^0.44.0",
     "morgan": "^1.10.0",
     "postcss-nested": "^5.0.6",
     "prettier": "2.4.1",
-    "react": "^17.0.2",
+    "react": "18.2.0",
     "react-chartjs-2": "^4.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "18.2.0",
     "react-router-dom": "^6.3.0",
     "sass": "^1.49.11",
     "source-map-explorer": "^2.5.2",
@@ -73,10 +80,14 @@
     "styletron-engine-atomic": "^1.4.8",
     "styletron-engine-monolithic": "^0.0.6",
     "styletron-react": "^6.0.2",
+    "swr": "^1.3.0",
     "typescript": "^4.5.4",
     "vite": "^2.9.8",
     "vite-plugin-pages": "^0.22.0",
     "vite-tsconfig-paths": "^3.3.17",
     "xnr": "^0.1.1"
+  },
+  "devDependencies": {
+    "http-proxy-middleware": "^2.0.6"
   }
 }
