@@ -1,7 +1,6 @@
 import { ReactNode } from "react";
 import { Item } from "baseui/side-navigation";
 import { ResponsiveNav } from "./ResponsiveNav";
-import { useLocation } from "react-router-dom";
 
 /* Empty Page (No Nav) */
 export const EmptyLayout = ({
@@ -31,7 +30,7 @@ export const EmptyLayout = ({
 };
 
 /* Responsive Nav Page */
-const MENU_ITEMS_ALL: Item[] = [
+const MENU_ITEMS: Item[] = [
   {
     title: "Your skills",
     itemId: "/skills",
@@ -45,63 +44,14 @@ const MENU_ITEMS_ALL: Item[] = [
         itemId: "/skills/languages",
       },
       {
-        title: "Skills you'd like to develop",
+        title: "Skills to develop",
         itemId: "/skills/develop",
       },
     ],
   },
   { title: "Your details", itemId: "/details" },
-  {
-    title: "Starter",
-    itemId: "/starter",
-    subNav: [
-      {
-        title: "Docs",
-        itemId: "/starter/docs",
-        subNav: [
-          {
-            title: "Directory Structure",
-            itemId: "/starter/docs/structure",
-          },
-          {
-            title: "How-To Mini Guides",
-            itemId: "/starter/docs/mini-guides",
-          },
-          {
-            title: "Components Guide",
-            itemId: "/starter/docs/components",
-          },
-          {
-            title: "CSS Utilities Guide",
-            itemId: "/starter/docs/css-guide",
-          },
-        ],
-      },
-      {
-        title: "Test Pages",
-        itemId: "/starter/test-pages",
-        subNav: [
-          {
-            title: "Basic Tests",
-            itemId: "/starter/test-pages/basics",
-          },
-          {
-            title: "Chart.js Test",
-            itemId: "/starter/test-pages/chart",
-          },
-          {
-            title: "CSS Defaults Test",
-            itemId: "/starter/test-pages/html-test",
-          },
-        ],
-      },
-    ],
-  },
+  { title: "[debug-menu]", itemId: "/debug" },
 ];
-
-const MENU_ITEMS: Item[] = MENU_ITEMS_ALL.filter(({ itemId }) => {
-  return !(itemId ?? "").startsWith("/starter");
-});
 
 export const MenuLayout = ({
   children,
@@ -110,13 +60,8 @@ export const MenuLayout = ({
   children?: ReactNode;
   maxWidth?: number;
 }) => {
-  const location = useLocation();
-
   return (
-    <ResponsiveNav
-      width={280}
-      items={location.pathname.startsWith("/starter") ? MENU_ITEMS_ALL : MENU_ITEMS}
-    >
+    <ResponsiveNav width={220} items={MENU_ITEMS}>
       <main className="mh-auto" style={{ maxWidth }}>
         <div className="rm-margin">{children}</div>
       </main>
