@@ -3,7 +3,7 @@ import { MenuLayout } from "_/components/Layouts";
 const SkillsToDevelop = () => {
   return (
     <MenuLayout>
-      <h1 className="D-S">Skills you'd like to develop</h1>
+      <h1 className="D-S">Skills to develop</h1>
     </MenuLayout>
   );
 };
