@@ -1,12 +1,39 @@
 import react from "@vitejs/plugin-react";
 import path from "node:path";
-import { defineConfig } from "vite";
+import { URL } from "node:url";
+import { defineConfig, ProxyOptions } from "vite";
 import pages from "vite-plugin-pages";
 import { parseHost } from "./.scripts/lib/url";
 import { getEnv } from "./.scripts/lib/project";
 
 const ENV = getEnv();
 const frontEndHost = parseHost(ENV.VITE_FRONTEND_HOST);
+
+let proxyHost: string | false = false;
+try {
+  let proxyRootUrl = ENV.VITE_PROXY_ROOT_URL;
+  while (proxyRootUrl.endsWith("/")) proxyRootUrl = proxyRootUrl.slice(0, -1);
+  let parsedUrl = new URL(proxyRootUrl).href;
+  while (parsedUrl.endsWith("/")) parsedUrl = parsedUrl.slice(0, -1);
+  proxyHost = parsedUrl;
+} catch {
+  //
+}
+
+let proxy: Record<string, string | ProxyOptions> | undefined;
+if (proxyHost !== false) {
+  proxy = {
+    "/api": {
+      target: proxyHost,
+      changeOrigin: true,
+      configure: (proxy) => {
+        proxy.on("proxyReq", (r) => {
+          r.path = r.path.replace("/api/", "/");
+        });
+      },
+    },
+  };
+}
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -16,9 +43,11 @@ export default defineConfig({
       ignored: ["**/tsconfig.json"],
     },
     port: frontEndHost.port,
+    proxy,
   },
   esbuild: {
     minifyIdentifiers: false,
+    logOverride: { "this-is-undefined-in-esm": "silent" },
   },
   resolve: {
     alias: [{ find: "_", replacement: path.resolve(process.cwd(), "src") }],
